# openapi.yaml
openapi: 3.0.3
info:
  title: MERGE STARS API
  version: 1.0.0
  description: >
    MERGE ⭐ STARS — API Specification v1.0 (Sprint 1 scope).
    Base URL is /api/v1. Responses use { data, meta } envelopes. Errors use { error, meta }.
servers:
  - url: http://localhost:3000/api/v1
    description: Local (Docker / dev)
...
paths:
...
  /audit/events:
    post:
      tags: [Audit]
      summary: Write audit event (best-effort)
      operationId: auditWrite
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuditWriteRequest'
      responses:
        '201':
          description: Written
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      tags: [Audit]
      summary: List audit events (admin)
      operationId: auditList
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Q'
        - name: action
          in: query
          required: false
          schema: { type: string }
        - name: actor_id
          in: query
          required: false
          schema: { type: string }
      responses:
        '200':
          description: List
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Q:
      name: q
      in: query
      required: false
      schema: { type: string }
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    Meta:
      type: object
      properties:
        request_id: { type: string }
    OkEnvelope:
      type: object
      properties:
        data:
          type: object
          properties:
            ok: { type: boolean, example: true }
        meta:
          $ref: '#/components/schemas/Meta'
...
    ErrorDetail:
      type: object
      properties:
        path: { type: string }
        reason: { type: string }
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details:
          type: array
          items: { $ref: '#/components/schemas/ErrorDetail' }
    ErrorEnvelope:
      type: object
      properties:
        error: { $ref: '#/components/schemas/ErrorBody' }
        meta: { $ref: '#/components/schemas/Meta' }

    # ── Auth ──────────────────────────────────────────────────────────────────
...
    AuditWriteRequest:
      type: object
      required: [action]
      properties:
        action: { type: string }
        object_type: { type: string, nullable: true }
        object_id: { type: string, nullable: true }
        meta:
          type: object
          additionalProperties: true
    AuditEvent:
      type: object
      properties:
        id: { type: string }
        actor_id: { type: string }
        actor_role: { type: string }
        action: { type: string }
        object_type: { type: string, nullable: true }
        object_id: { type: string, nullable: true }
        result: { type: string, enum: [SUCCESS, FAILURE] }
        ip_address: { type: string, nullable: true }
        meta:
          type: object
          additionalProperties: true
        created_at: { type: string, format: date-time }
    PaginatedAudit:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/AuditEvent' }
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
    AuditListEnvelope:
      type: object
      properties:
        data: { $ref: '#/components/schemas/PaginatedAudit' }
        meta: { $ref: '#/components/schemas/Meta' }

    # ── Uploads ───────────────────────────────────────────────────────────────
...